+++ test/bigtest/mocks/index.js
@@ -2,3 +2,4 @@ export * from './job-executions';
 export * from './jobs-logs';
 export * from './job-log';
 export * from './associated-job-profiles';
+export * from './associated-mapping-profile';