+++ test/bigtest/network/scenarios/fetch-action-profiles-success.js
@@ -1,3 +1,5 @@
+import { associatedJobProfiles } from '../../mocks';
+
 export default server => {
   server.create('action-profile', {
     action: 'CREATE',
@@ -33,6 +35,8 @@ export default server => {
   });
 
   server.get('/data-import-profiles/actionProfiles');
+  server.get('/data-import-profiles/actionProfiles/:id');
+  server.get('/data-import-profiles/profileAssociations/:id/masters', associatedJobProfiles);
   server.post('/data-import-profiles/actionProfiles', (_, request) => {
     const params = JSON.parse(request.requestBody);
     const record = server.create('match-profile', params);